Placed component in separate file and moved prop further up the chain.
